requirements.txt
python-jenkins>=1.0.0
google-generativeai>=0.3.0
typer>=0.9.0
rich>=13.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
